Working captiv8 plugin from config.ini to fully rendered graph
